+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change Log
 
-## *WIP*
+## [0.11.0]
+* refactor: improve *JSDoc* support (`ssh.js`)
 * fix: throw an error when no value was provided for a validator which can be defined multiple times (`arg.js`)
 * feat: possibility to override the usage message for enum values (`arg.js`)
+* feat: new _bootstrap_ script to create a new project
+* fix: detect whether or not query parameters exist in the url (`curl.js`)
 
 ## [0.10.0]
 * feat: improve cmdline arguments parsing (`arg.js`)